refactor(cordova): move GenericAd to shared
